ATM: I was passing my Double variable to an Integer. Fix it back to Double AtmMessages: I fixed one of the Error messages Verification: Removed static from fourthInteraction pseudocode: I structured more like how ATM looks

+++ src/main/java/au/com/nuvento/atm/ATM.java
@@ -53,7 +53,7 @@ public class ATM
 
         // Showing Balance and enter amount
         interactions.enterInAccount(userName, options, userOption);
-        double userAmount = Integer.parseInt(sc.nextLine());
+        double userAmount = Double.parseDouble(sc.nextLine());
         double newBal = commands.newBalance(userName, options, userOption, userAmount);
         verification.fourthInteraction(userName, options, newBal, userAmount);
         System.out.println("Your new Balance is: " + newBal);
